♻️ refactor:  themeをconfigに移動
#3
